fix: removed nonexisted methods
